Refactor mosaic_fsdp.py
