+++ composer/trainer/dist_strategy.py
@@ -218,7 +218,8 @@ def prepare_fsdp_module(
     if not is_torch_2_0:
         from torch.distributed.fsdp.flatten_params_wrapper import FlattenParamsWrapper
 
-    from composer.trainer.mosaic_fsdp import backward_prefetch_map, get_cpu_offload, get_mixed_precision, sharding_map
+    from composer.trainer.mosaic_fsdp_utils import (BACKWARD_PREFETCH_MAP, SHARDING_MAP, get_cpu_offload,
+                                                    get_mixed_precision)
 
     set_fsdp_default(fsdp_config)
 
@@ -310,7 +311,7 @@ def prepare_fsdp_module(
         optim.state.clear()
 
     sharding_map_key = fsdp_config['sharding_strategy'].upper()
-    sharding_strategy = sharding_map[sharding_map_key]
+    sharding_strategy = SHARDING_MAP[sharding_map_key]
 
     cpu_offload = get_cpu_offload(cpu_offload=fsdp_config['cpu_offload'])
 
@@ -340,7 +341,7 @@ def prepare_fsdp_module(
                 f'Consider using `amp` or `bf16` for precision or setting param_dtype in mixed_precision to `None` '
                 f'with sharding strategy `{sharding_map_key}.`')
 
-    backward_prefetch = backward_prefetch_map[fsdp_config['backward_prefetch'].upper()]
+    backward_prefetch = BACKWARD_PREFETCH_MAP[fsdp_config['backward_prefetch'].upper()]
     activation_checkpointing = fsdp_config['activation_checkpointing']
     activation_cpu_offload = fsdp_config['activation_cpu_offload']
     sync_module_states = fsdp_config['sync_module_states']
